update based upon feedback
